Bringing the power of sized types to MSVC compilers. This is my attempt to fix the underlying issue resulting in r16229 and r16232. I fully admit this is a bit out of my league so please do review.
git-svn-id: https://simulationcraft.googlecode.com/svn/trunk@16233 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ engine/sc_reforge_plot.cpp
@@ -213,7 +213,7 @@ void reforge_plot_t::analyze_stats()
     if ( sim -> report_progress )
     {
       tmp += ":";
-      int n_spaces = 23 - tmp.length();
+      ssize_t n_spaces = 23 - tmp.length();
       if ( n_spaces > 0 )
         tmp.append( n_spaces, ' ' );
     }
